+++ apps/home/src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import { useTranslation } from 'react-i18next';
+import { useTranslation } from '../i18n';
 import { SurfaceCard } from 'corona/surface';
 
 import { useDocumentHead } from '../hooks/useDocumentHead';